[Fix]: 패키지 구조 변경 및 전역 ExceptionHandler 생성, 빈 폴더에 대한 체킹 추가
