--- hanbatbox-api/src/main/java/com/hbu/hanbatbox/S3/service/S3Service.java
@@ -1,37 +1,0 @@
-package com.hbu.hanbatbox.S3.service;
-
-import java.nio.file.Path;
-import java.util.HashMap;
-import java.util.Map;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-import software.amazon.awssdk.core.sync.RequestBody;
-import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.PutObjectRequest;
-
-@Service
-public class S3Service {
-
-  private final S3Client s3Client;
-
-  @Value("${spring.cloud.aws.S3.bucket}")
-  private String bucketName;
-
-  @Autowired
-  public S3Service(S3Client s3Client) {
-    this.s3Client = s3Client;
-  }
-
-  public void uploadFile(String objectKey, String title, String password, Path filePath) {
-    Map<String, String> metadata = new HashMap<>();
-    metadata.put("title", title);
-    metadata.put("password", password);
-
-    PutObjectRequest putObjectRequest = PutObjectRequest.builder().bucket(bucketName).key(objectKey)
-        .metadata(metadata).build();
-
-    s3Client.putObject(putObjectRequest, RequestBody.fromFile(filePath));
-    System.out.println("File uploaded successfully to " + bucketName + "/" + objectKey);
-  }
-}
