+++ hanbatbox-api/src/main/java/com/hbu/hanbatbox/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.hbu.hanbatbox.S3;
+package com.hbu.hanbatbox.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
